trading_ai/requirements.txt
# AI Trading Signal System - Dependencies
# Python 3.10+ required

# ============================================================================
# DATA & MACHINE LEARNING
# ============================================================================
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.15.0
optuna>=3.5.0

# ============================================================================
# MARKET DATA SOURCES
# ============================================================================
yfinance>=0.2.33
kagglehub>=0.2.0
fredapi>=0.5.1
newsapi-python>=0.2.7
alpaca-trade-api>=3.0.0

# ============================================================================
# TECHNICAL ANALYSIS
# ============================================================================
ta>=0.11.0
scipy>=1.11.0

# ============================================================================
# API & ASYNC
# ============================================================================
anthropic>=0.18.0
requests>=2.31.0
httpx>=0.26.0
aiohttp>=3.9.0

# ============================================================================
# CONFIGURATION & UTILITIES
# ============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0.1
joblib>=1.3.0

# ============================================================================
# VISUALIZATION & DASHBOARD
# ============================================================================
streamlit>=1.30.0
plotly>=5.18.0

# ============================================================================
# TESTING & DEVELOPMENT
# ============================================================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
responses>=0.24.0

# ============================================================================
# OPTIONAL - Additional useful packages
# ============================================================================
# openai>=1.10.0  # Alternative LLM if needed
# beautifulsoup4>=4.12.0  # Web scraping if needed
# selenium>=4.16.0  # Browser automation if needed
